feat: generate full combined rss
This is a major re-write to how we deal with rss. Everything is
configured via toml and the command `pnpm start` will now generate a
combined rss file for all the content in the `[[feeds]]` section.

**Note**
It doesn't limit the final rss file that gets generated - the total
items will be the sum of all the teims in the feeds configuration/